requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
blessed==1.20.0
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
cchardet @ file:///C:/b/abs_63f5b8zxp9/croot/cchardet_1683054562010/work
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet @ file:///C:/ci/chardet_1607706937985/work
charset-normalizer==3.1.0
click==8.1.3
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
commonmark @ file:///Users/ktietz/demo/mc3/conda-bld/commonmark_1630649545323/work
contourpy==1.0.7
croniter==1.3.14
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
Cython @ file:///C:/b/abs_6ev7kndkxh/croot/cython_1676559821329/work
dateutils==0.6.12
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepdiff==6.3.0
Distance==0.1.3
dnspython==2.3.0
docopt==0.6.2
editdistance @ file:///C:/b/abs_83s9pyo0en/croot/editdistance_1670956756568/work
email-validator==1.3.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.7.3
fastapi==0.88.0
ffmpy==0.3.0
filelock==3.10.3
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.5.0
future @ file:///C:/b/abs_3dcibf18zi/croot/future_1677599891380/work
google-auth==2.16.3
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.13.4
hydra-core==1.3.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
ijson==3.2.0.post0
importlib-metadata==6.1.0
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
inflect @ file:///C:/ci/inflect_1614797534456/work
inquirer==3.1.3
ipadic==1.0.0
ipython @ file:///C:/b/abs_d3h279dv3h/croot/ipython_1676582236558/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
jieba==0.42.1
Jinja2==3.1.2
jinxed==1.2.0
jiwer==3.0.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0.post2
lightning==1.9.0
lightning-cloud==0.5.34
lightning-utilities==0.8.0
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mecab-python3==1.0.5
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-toolkit==1.17.0
networkx==3.0
numba==0.56.4
numexpr @ file:///C:/b/abs_afm0oewmmt/croot/numexpr_1683221839116/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682981337988/work
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.1
OpenCC==1.1.1
ordered-set==4.1.0
packaging==22.0
pandas @ file:///C:/b/abs_b9kefbuby2/croot/pandas_1677835593760/work
pangu==4.0.6.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pipreqs==0.4.13
plac==1.3.5
pooch==1.6.0
progress==1.6
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.3
psutil==5.9.5
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyannote.core==5.0.0
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
pypinyin==0.48.0
pypinyin-dict==0.5.0
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytest-runner==6.0.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.9.0
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz @ file:///C:/b/abs_3es0o30520/croot/rapidfuzz_1679593472486/work
readchar==4.0.5
regex==2023.3.23
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.95
sentry-sdk==1.19.1
setproctitle==1.3.2
shellingham==1.5.0.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.4.1
soxr==0.3.4
sphfile==1.0.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.22.0
starsessions==1.3.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
text-unidecode==1.3
textdistance==4.5.0
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
transformers==4.27.4
typed-ast==1.5.4
typer==0.7.0
typing_extensions @ file:///C:/b/abs_a1bb332wcs/croot/typing_extensions_1681939523095/work
uc-micro-py==1.0.1
ujson==5.7.0
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdataset==0.1.62
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
win32-setctime==1.1.0
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
youtokentome==1.0.6
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
